tests: extract resilience and error-path cases
