+++ tests/test_harness.h
@@ -5,6 +5,7 @@
 
 extern int status;
 extern tealet_t *g_main;
+extern int talloc_fail;
 
 typedef struct lock_snapshot_t {
   int lock_calls;
@@ -13,6 +14,7 @@ typedef struct lock_snapshot_t {
 
 void init_test(void);
 void fini_test(void);
+void init_test_extra(tealet_alloc_t *alloc, size_t extrasize);
 
 void lock_snapshot_take(lock_snapshot_t *snap);
 void lock_snapshot_assert_delta_one(const lock_snapshot_t *before);
